getId() of ZigBeeEndpoint returns a short value, and device version in ZigBeeSimpleDescriptor is linked to a byte value

+++ org.osgi.service.zigbee/src/org/osgi/service/zigbee/ZigBeeEndpoint.java
@@ -95,7 +95,7 @@ public interface ZigBeeEndpoint {
 	 * @return identifier of the endpoint represented by this object, value
 	 *         ranges from 1 to 240.
 	 */
-	public byte getId();
+	public short getId();
 
 	/**
 	 * @return The IEEE Address of the node containing this
